tlp: allow cpu turbo on battery as well
It is conceivable, and is in fact claimed by the kernel devs that not
manually locking down the clock is better for power saving, as the
kernel may choose to clock up to finish some work faster and thus
quickly return to the extreme low power states. Basically, this is the
philosophy of bang-bang control.

Now, these claims have been disputed, in particular the devs don't
really know what they are doing wrt cpu power saving since their claim that
"ondemand" (default) is the best governor regardless of ac/bat is completely bogus,
and has been bogus for over 2 years now. In particular "powersave" yields savings.

However, this gives them the benefit of doubt. This may be reverted if I
can see concrete drawbacks, hopefully there won't be.