vim: remove gutentags
gutentags still has the issue of spinning up a ctags process that locks
down a CPU, causing annoying fan issues. This happens on occassion, not
always, and has not been fixed.

Lack the patience to debug it and/or submit issue, as and when I need tags again
they will be manually generated.