sakura: switch to sakura from lxterminal
lxterminal was behaving very poorly with the port to Awesome 4.0.
Basically, changing the screen size was causing it to not respond to
keyboard input. No real idea why this happens: all I can say is that
1. This worked fine with Awesome 3.5.
2. lxterminal was the only culprit on my system.

Lack the energy to investigate this, and I don't have strong loyalty to
any particular terminal emulator. As such, I simply moved to sakura.