mpv: default to opengl-hq
Finally measured the battery life more exhaustively. The difference is
on the order of just a few percent; note that in such cases there are
far more important things to worry about, such as Google/Youtube's
nonsensical insistence on vp9 in spite of the lack of efficient hardware
decoders on most systems.